requirements.txt
Django==3.2.7
djangorestframework==3.13.1
djangorestframework_simplejwt==5.0.0
